+++ internal/resources/resource_role.go
@@ -200,7 +200,7 @@ func (r *RoleResource) Update(ctx context.Context, req resource.UpdateRequest, r
 	})
 }
 
-func (r *RoleResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *RoleResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) { //nolint:dupl
 	var state Role
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {